requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
click==8.1.7
croniter==3.0.3
Flask==3.0.3
Flask-APScheduler==1.13.1
flask-babel==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
RPi.GPIO==0.7.1
six==1.16.0
spidev==3.6
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4